removed tweet_posts/instagram_posts controllers, moved methods to feeds controller since we weren't generating views for these controllers and so we could auto refresh for the index
