--- app/helpers/instagram_posts_helper.rb
@@ -1,2 +1,0 @@
-module InstagramPostsHelper
-end
