+++ config/routes.rb
@@ -20,7 +20,5 @@ Rails.application.routes.draw do
 
   resources :instagrams, only: [:create, :destroy]
   resources :tweets, only: [:create, :destroy]
-  resources :instagram_posts, only: [:create]
-  resources :tweet_posts, only: [:create]
 
 end
